requirements.txt
tekore
google-cloud-bigquery
